Updated avalonia to 11.0.6. Screen recorded setting up write formats for vs code but got side tracked setting up dump creation and posting an issue about the weird gpu freeze bug that happens. I think I fixed issue logging to files, noticed it wasn't recording exceptions but now I think I fixed it. Finding quite a few little bugs...

+++ MonkeyPaste.Avalonia/Controls/Content/CefNetWebView/Cef/MpAvCefNetApplication.cs
@@ -254,7 +254,7 @@ namespace MonkeyPaste.Avalonia {
 
         public static void ResetCefNetLogging() {
 
-            if (LogFilePath.IsFile()) {
+            if (LogFilePath.IsFile() && !MpFileIo.IsFileInUse(LogFilePath)) {
                 MpFileIo.DeleteFile(LogFilePath);
             }
         }
